requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
geotext==0.4.0
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
